refactor: View 클래스 분리
